refactor: WASM bindings modularization + CI pipeline overhaul
Major architectural refactoring of the WASM bindings layer:
- Extract event handlers, selection, rendering, and helpers into separate modules
- Add missing selection module declaration (selectAll, copySelection, etc.)
- Convert struct fields to pub(crate) for cross-module access
- Remove unused events.rs placeholder structs

CI pipeline improvements:
- Decompose monolithic job into focused jobs: fmt, clippy, rust, security, deny, wasm, e2e
- Add cargo-audit and cargo-deny checks
- Add concurrency settings to cancel in-progress runs
- Add workflow_dispatch trigger
- Add permissions hardening

Other changes:
- Update rustfmt.toml with use_field_init_shorthand and use_try_shorthand
- Add clippy.toml thresholds for cognitive complexity and argument count
- Add deny.toml for supply chain security
- Add quality-gates.sh script for local CI
- Improve E2E tests with polling-based waits instead of arbitrary delays
- Add cargo profile optimizations for faster dev builds
