+++ .cargo/config.toml
@@ -1,3 +1,18 @@
 [target.wasm32-unknown-unknown]
 rustflags = ["-C", "target-feature=+simd128", "-C", "target-feature=+bulk-memory"]
 runner = "wasm-bindgen-test-runner"
+
+[profile.dev]
+debug = "line-tables-only"
+
+[profile.dev.package."*"]
+debug = false
+
+[profile.dev.build-override]
+opt-level = 3
+
+[alias]
+check-all = "check --workspace --all-features"
+test-all = "nextest run --workspace"
+fmt-check = "fmt --all -- --check"
+lint = "clippy --workspace --all-targets --all-features -- -D warnings"
