+++ .github/workflows/ci.yml
@@ -4,15 +4,78 @@ on:
   push:
     branches: [main, develop]
   pull_request:
-    branches: [main]
+    branches: [main, develop]
+  workflow_dispatch:
+
+permissions:
+  contents: read
 
 env:
   CARGO_TERM_COLOR: always
   RUST_BACKTRACE: 1
+  CARGO_REGISTRIES_CRATES_IO_PROTOCOL: sparse
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
 
 jobs:
+  changes:
+    name: Detect Changes
+    runs-on: ubuntu-latest
+    outputs:
+      code: ${{ steps.filter.outputs.code }}
+    steps:
+      - uses: actions/checkout@v6
+      - uses: dorny/paths-filter@v3
+        id: filter
+        with:
+          filters: |
+            code:
+              - 'src/**'
+              - 'tests/**'
+              - 'benches/**'
+              - 'Cargo.toml'
+              - 'Cargo.lock'
+              - 'rust-toolchain.toml'
+
+  fmt:
+    name: Format
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          components: rustfmt
+      - name: Run rustfmt
+        run: cargo fmt --all -- --check
+
+  clippy:
+    name: Clippy
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+      - uses: dtolnay/rust-toolchain@stable
+        with:
+          components: clippy
+      - uses: Swatinem/rust-cache@v2
+      - name: Clippy (all targets, all features)
+        run: cargo clippy --workspace --all-targets --all-features -- -D warnings
+
   rust:
-    name: Rust (wasm-pack + Tests)
+    name: Rust (Build + Tests)
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
     runs-on: ubuntu-latest
     timeout-minutes: 15
 
@@ -31,21 +94,8 @@ jobs:
         with:
           workspaces: ./->target
 
-      - name: Check formatting
-        run: cargo fmt --all -- --check
-
-      - name: Run clippy
-        run: cargo clippy --all-targets --all-features -- -D warnings
-
-      - name: Setup pnpm
-        uses: pnpm/action-setup@ac6db6d3c1f721f886538a378a2d73e85697340a # v6
-        with:
-          version: 10
-
-      - name: Build WASM
-        run: |
-          npm install -g wasm-opt
-          pnpm run build:wasm
+      - name: Build
+        run: cargo build --all-targets
 
       - name: Run tests (Native)
         run: cargo test --all
@@ -55,16 +105,80 @@ jobs:
         env:
           WASM_BINDGEN_TEST_ONLY_NODE: 1
 
-      - name: Upload WASM
-        uses: actions/upload-artifact@v7
+  security:
+    name: Security Audit
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+      - uses: dtolnay/rust-toolchain@stable
+      - uses: Swatinem/rust-cache@v2
+      - name: Install cargo-audit
+        uses: taiki-e/install-action@v2
         with:
-          name: wasm pkg
+          tool: cargo-audit
+      - name: Generate lockfile
+        run: cargo generate-lockfile
+      - name: Run cargo audit
+        run: cargo audit
+
+  deny:
+    name: Dependency Policy (cargo-deny)
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v6
+      - uses: EmbarkStudios/cargo-deny-action@v2
+
+  wasm:
+    name: WASM Build
+    needs: [changes]
+    if: >
+      github.event_name != 'pull_request' ||
+      needs.changes.outputs.code == 'true'
+    runs-on: ubuntu-latest
+    timeout-minutes: 15
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v6
+
+      - name: Setup mise
+        uses: jdx/mise-action@v4
+
+      - name: Add WASM target
+        run: rustup target add wasm32-unknown-unknown
+
+      - name: Cache cargo
+        uses: Swatinem/rust-cache@v2
+        with:
+          workspaces: ./->target
+
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v6
+        with:
+          version: 10
+
+      - name: Build WASM
+        run: |
+          npm install -g wasm-opt
+          pnpm run build:wasm
+
+      - name: Upload WASM
+        uses: actions/upload-artifact@v4
+        with:
+          name: wasm-pkg
           path: web/pkg
 
   e2e:
     name: E2E Tests (Playwright)
     runs-on: ubuntu-latest
-    needs: rust
+    needs: [rust, wasm]
     timeout-minutes: 20
 
     steps:
@@ -72,17 +186,17 @@ jobs:
         uses: actions/checkout@v6
 
       - name: Setup Node.js
-        uses: actions/setup-node@v6
+        uses: actions/setup-node@v4
         with:
           node-version: '22'
 
       - name: Setup pnpm
-        uses: pnpm/action-setup@ac6db6d3c1f721f886538a378a2d73e85697340a # v6
+        uses: pnpm/action-setup@v6
         with:
           version: 10
 
       - name: Cache pnpm
-        uses: actions/cache@v5
+        uses: actions/cache@v4
         with:
           path: ~/.local/share/pnpm/store
           key: ${{ runner.os }}-node-${{ hashFiles('**/pnpm-lock.yaml') }}
@@ -98,9 +212,9 @@ jobs:
         run: npx playwright install chromium
 
       - name: Download WASM
-        uses: actions/download-artifact@v8
+        uses: actions/download-artifact@v4
         with:
-          name: wasm pkg
+          name: wasm-pkg
           path: web/pkg
 
       - name: Run E2E tests
@@ -119,3 +233,28 @@ jobs:
           npx playwright test --project=chromium --timeout=90000
         env:
           CI: true
+
+  ci-success:
+    name: CI Success
+    needs: [fmt, clippy, rust, security, deny, wasm, e2e]
+    runs-on: ubuntu-latest
+    if: always()
+    steps:
+      - name: Check all jobs passed
+        run: |
+          results=(
+            "${{ needs.fmt.result }}"
+            "${{ needs.clippy.result }}"
+            "${{ needs.rust.result }}"
+            "${{ needs.security.result }}"
+            "${{ needs.deny.result }}"
+            "${{ needs.wasm.result }}"
+            "${{ needs.e2e.result }}"
+          )
+          for res in "${results[@]}"; do
+            if [[ "$res" == "failure" || "$res" == "cancelled" ]]; then
+              echo "One or more critical CI jobs failed or were cancelled"
+              exit 1
+            fi
+          done
+          echo "All CI jobs passed or were correctly skipped!"
