+++ rustfmt.toml
@@ -1,3 +1,5 @@
 edition = "2021"
 max_width = 100
-use_small_heuristics = "Default"
+use_field_init_shorthand = true
+use_try_shorthand = true
+newline_style = "Unix"
